src/lmgtfu-agent/requirements.txt
python-dotenv
pyautogen
autogenstudio
pytest
langchain
langchain_community
langchain-openai
bs4